The to_prepare block should be done in the Engine
This will make it run before other gems that are declared after spree_mobility in the Gemfile. If done in an initializer
like before, then this would not be the case. This replaces the temporary fix for PaymentMethod STI issues - as long as
spree_mobility is before other PaymentMethod gems in the Gemfile, the new payment methods will inherit the translations
association and settings from PaymentMethod.

Fixes #7